Load session and auth parameters via CLI
The OpenStack keystoneauth module is the current way to load these
values, and it (should) be more future-proof, as it will automatically
gain the ability to parse new options in the future.

There isn't a great way to load the Adapter options yet, unfortunately -
that will probably require an upstream patch.