Remove Adapter session due to lack of client support
Some clients, notably the neutronclient and probably others, do not
support the automatic version discovery that is done when you use an
Adapter as the Session interface. The Adapter provides some
automatically set kwargs on all the request calls, and one of those is
'endpoint_filter'. When this is set, endpoint discovery automatically
occurs, and requests to the service API are automatically versioned. The
neutronclient prefixes the version on all calls itself, meaning the
version ends up prefixed twice, causing problems.

The solution for this, while preserving the region functionality, is to
not use the Adapter but pass the region_name in on the client
construction calls, as was done prior.

When all the clients support this version discovery this can go away.