Add check for Gnocchi measurements
This adds a Gnocchi check that just looks at the total number of
measurements found for a metric name. This is written for the use-case
of knowing if Gnocchi has a problem delivering metrics or not.
Monitoring metric values is best left to better reporting and
visualization tools, like Grafana.